refactor: chuyển đổi vị trí file RoleUserText
