+++ app/src/main/java/com/pwhs/quickmem/presentation/app/library/classes/component/ClassItem.kt
@@ -34,7 +34,7 @@ import coil.compose.AsyncImage
 import com.pwhs.quickmem.R
 import com.pwhs.quickmem.domain.model.classes.GetClassByOwnerResponseModel
 import com.pwhs.quickmem.domain.model.users.UserResponseModel
-import com.pwhs.quickmem.presentation.app.profile.component.RoleUserText
+import com.pwhs.quickmem.presentation.component.RoleUserText
 import com.pwhs.quickmem.ui.theme.QuickMemTheme
 
 @Composable
