+++ app/src/main/java/com/pwhs/quickmem/presentation/app/library/study_set/component/StudySetItem.kt
@@ -39,7 +39,7 @@ import com.pwhs.quickmem.domain.model.color.ColorModel
 import com.pwhs.quickmem.domain.model.study_set.GetStudySetResponseModel
 import com.pwhs.quickmem.domain.model.subject.SubjectModel
 import com.pwhs.quickmem.domain.model.users.UserResponseModel
-import com.pwhs.quickmem.presentation.app.profile.component.RoleUserText
+import com.pwhs.quickmem.presentation.component.RoleUserText
 import com.pwhs.quickmem.ui.theme.QuickMemTheme
 import com.pwhs.quickmem.util.toColor
